More integration test reorganization
