+++ tests/Tgstation.Server.Tests/Live/Instance/ConcreteHttpClientFactory.cs
@@ -1,7 +1,7 @@
 ﻿using Tgstation.Server.Common;
 using Tgstation.Server.Host.Core;
 
-namespace Tgstation.Server.Tests
+namespace Tgstation.Server.Tests.Live.Instance
 {
 	sealed class ConcreteHttpClientFactory : IAbstractHttpClientFactory
 	{
